# apps/api/src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VideoQ API",
    "version": "1.0.0",
    "description": "Hono Worker OpenAPI surface (Django Admin / DRF spectacular 置換)."
  },
  "paths": {
...
    "/api/videos/groups/": {
      "get": { "summary": "List groups", "tags": ["videos"] },
      "post": {
        "summary": "Create group",
        "tags": ["videos"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ops/users/": {
      "get": {
        "summary": "List users (ops / superuser)",
        "tags": ["ops"],
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" } }
        ]
      }
    },
...
  }
}